Fix bugs in the routing in sample APIs to allow the endpoitns to be hit.
1) Modify web.config to allow PUT and DELETE to be hit. MVC by default only allows GET and POST.

2) Remove the front slash '/' and the query param from the attribute-based routing since they are prohibited.

3) Register the MVC attribute-based routing in RouteConfig.cs

4) Rename the controllers. MVC requires the controllers to end with suffix `Controller`. If we don't do this, we cannot hit the endpoints :(

5) Fix wrong parameters in SampleV4Controller.
