+++ test/Microsoft.OpenApi.CSharpAnnotations.DocumentGeneration.Tests.SampleApis/App_Start/RouteConfig.cs
@@ -24,6 +24,8 @@ namespace Microsoft.OpenApi.CSharpAnnotations.DocumentGeneration.Tests.SampleApi
         public static void RegisterRoutes(RouteCollection routes)
         {
             routes.IgnoreRoute("{resource}.axd/{*pathInfo}");
+
+            routes.MapMvcAttributeRoutes();
         }
     }
 }
